Also rename the files
